fix: ajusted delete mentoring button at Member areas

+++ src/layouts/mentor/area-de-membros/index.tsx
@@ -1,15 +1,18 @@
 import { FC, useState } from 'react';
-import { DeleteForever } from '@mui/icons-material';
 import Box from '@mui/material/Box';
 import Divider from '@mui/material/Divider';
+import InputLabel from '@mui/material/InputLabel';
+import MenuItem from '@mui/material/MenuItem';
+import Select from '@mui/material/Select';
 import { useTheme } from '@mui/material/styles';
 import Typography from '@mui/material/Typography';
 import { withPageAuth } from '@supabase/auth-helpers-nextjs';
 import Image from 'next/future/image';
 import { useRouter } from 'next/router';
+import { toast } from 'react-toastify';
 import ContentWidthLimit from '~/components/modules/ContentWidthLimit';
 import ModalComponent from '~/components/modules/Modal';
-import { PublicRoutes } from '~/consts';
+import { MentorRoutes, PublicRoutes } from '~/consts';
 import { useProducts } from '~/hooks/useProducts';
 import { GetProfile } from '~/services/profile.service';
 import CreateProductDialog from './components/CreateProductDialog';
@@ -19,10 +22,12 @@ import {
   AbsoluteTopBox,
   AreaWrapper,
   CollorFullMentorfy,
-  CreatAreaButton,
+  CreateAreaButton,
+  DeleteAreaButton,
   EmptyBox,
   ImageButton,
   ProductTitle,
+  ProductsSelectField,
 } from './styles';
 import PlusSvg from '~/../public/svgs/plus';
 
@@ -33,6 +38,7 @@ const MembersArea: FC<PageTypes.Props> = ({ user }) => {
   const [openCreatePage, setOpenCreatePage] = useState(false);
   const [showConfirmDelete, setShowConfirmDelete] = useState(false);
   const [productId, setProductId] = useState('');
+  const [open, setOpen] = useState(false);
 
   const handleDeleteProduct = async (id: string) => {
     setProductId(id);
@@ -40,11 +46,18 @@ const MembersArea: FC<PageTypes.Props> = ({ user }) => {
   };
 
   const confirmDeleteProduct = async () => {
-    setShowConfirmDelete(false);
-    await fetch(`/api/products?id=${productId}`, {
-      method: 'DELETE',
-    });
-    mutate();
+
+    if (!showConfirmDelete) {
+      setShowConfirmDelete(true);
+    } else {
+      await fetch(`/api/products?id=${productId}`, {
+        method: 'DELETE',
+      });
+      toast.success('Alterações salvas com sucesso');
+      mutate();
+      setOpen(false);
+      setShowConfirmDelete(false);
+    }
   };
 
   return (
@@ -54,11 +67,21 @@ const MembersArea: FC<PageTypes.Props> = ({ user }) => {
           <Box sx={{ float: 'left' }}>
             <Typography>Minhas Mentorias</Typography>
           </Box>
-          <Box sx={{ float: 'right' }} onClick={() => setOpenCreatePage(true)}>
-            <CreatAreaButton variant="outlined">
+          <Box sx={{ float: 'right' }} >
+            <CreateAreaButton variant="outlined" onClick={() => setOpenCreatePage(true)}>
               <PlusSvg fill={theme.palette.accent.main} />
               Criar Mentoria
-            </CreatAreaButton>
+            </CreateAreaButton>
+            <DeleteAreaButton
+              className="delete-icon"
+              onClick={(e) => {
+                setOpen(true);
+                // e.stopPropagation();
+                // handleDeleteProduct(area.id);
+              }}
+            >
+              Excluir Mentoria
+            </DeleteAreaButton>
           </Box>
         </Box>
         <Box
@@ -74,7 +97,7 @@ const MembersArea: FC<PageTypes.Props> = ({ user }) => {
           {products?.map((area, index) => (
             <AreaWrapper
               onClick={
-                () => {}
+                () => { router.push(MentorRoutes.members_area_editar + '/' + area.id); }
                 // router.push(MentorRoutes.members_area_editar + '/' + area.id)
               }
               key={index}
@@ -112,22 +135,6 @@ const MembersArea: FC<PageTypes.Props> = ({ user }) => {
                   <ProductTitle>{area?.title}</ProductTitle>
                 </AbsoluteBottomBox>
               )}
-              {(area as any).relations?.length === 0 && (
-                <DeleteForever
-                  className="delete-icon"
-                  color="error"
-                  style={{
-                    position: 'absolute',
-                    bottom: '15px',
-                    right: '15px',
-                    opacity: 0.5,
-                  }}
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    handleDeleteProduct(area.id);
-                  }}
-                />
-              )}
             </AreaWrapper>
           ))}
         </Box>
@@ -160,17 +167,48 @@ const MembersArea: FC<PageTypes.Props> = ({ user }) => {
       <CreateProductDialog open={openCreatePage} setOpen={setOpenCreatePage} />
       <ModalComponent
         onDelete={() => confirmDeleteProduct()}
-        onSave={() => setShowConfirmDelete(false)}
-        open={showConfirmDelete}
+        onSave={() => {
+          setOpen(false);
+          setShowConfirmDelete(false);
+        }
+        }
+        open={open}
+        setOpen={setOpen}
         title="Remover Mentoria"
         deleteMessage={true}
       >
-        <Box sx={{ textAlign: 'center' }}>
-          <DeleteText>
-            Ao remover, você perderá todos os dados relacionados a essa
-            mentoria. Deseja continuar mesmo assim?
-          </DeleteText>
-        </Box>
+        {showConfirmDelete ?
+          <Box sx={{ textAlign: 'center' }}>
+            <DeleteText>
+              Ao remover, você perderá todos os dados relacionados a essa
+              mentoria. Deseja continuar mesmo assim?
+            </DeleteText>
+          </Box>
+          :
+          <Box >
+            <DeleteText>Escolha qual produto deseja excluir</DeleteText>
+            <ProductsSelectField
+              required
+              sx={{ width: '100%', margin: '1rem 0' }}
+            >
+              <InputLabel shrink={true}>Produtos</InputLabel>
+              <Select
+                placeholder="Área de Membros Mentrofy"
+                label="Produtos"
+                color="secondary"
+                notched={true}
+              >
+                {products.map((product) => (
+                  <MenuItem key={product.id} value={product.id}
+                  onClick={() => setProductId(product.id)}>
+                    {product.title}
+                  </MenuItem>
+                ))}
+              </Select>
+            </ProductsSelectField>
+          </Box>
+        }
+
       </ModalComponent>
     </>
   );
